[TNT-181] move: Image 컴포넌트 파일 위치 수정
